Refactor: Introduce ExpandableFab composable
This commit introduces a new `ExpandableFab` composable to encapsulate the logic and UI for the expandable Floating Action Button menu.

Key changes:
- Created `ExpandableFab.kt` containing the new `ExpandableFab` composable.
- Migrated the FAB menu implementation from `DeckPickerScreen.kt` and `AnkiDroidApp.kt` to the new `ExpandableFab` composable.
- Updated `DeckPickerScreen.kt` and `AnkiDroidApp.kt` to use the `ExpandableFab` composable.

This refactoring improves code organization and reusability by centralizing the FAB menu logic into a dedicated composable.
