+++ AnkiDroid/src/main/java/com/ichi2/anki/ui/compose/DeckPickerScreen.kt
@@ -17,19 +17,14 @@
  ****************************************************************************************/
 package com.ichi2.anki.ui.compose
 
-import androidx.activity.compose.BackHandler
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.tween
 import androidx.compose.animation.expandVertically
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.shrinkVertically
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -44,17 +39,13 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Download
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.FloatingActionButtonMenu
-import androidx.compose.material3.FloatingActionButtonMenuItem
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.LargeTopAppBar
@@ -67,18 +58,13 @@ import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
-import androidx.compose.material3.ToggleFloatingActionButton
-import androidx.compose.material3.ToggleFloatingActionButtonDefaults.animateIcon
 import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.animateFloatingActionButton
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -86,21 +72,14 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Outline
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.asComposePath
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.graphics.vector.rememberVectorPainter
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.semantics.stateDescription
-import androidx.compose.ui.semantics.traversalIndex
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.LayoutDirection
@@ -313,203 +292,103 @@ fun DeckPickerScreen(
     searchFocusRequester: FocusRequester = FocusRequester(),
     snackbarHostState: SnackbarHostState = remember { SnackbarHostState() },
 ) {
-    var fabMenuExpanded by rememberSaveable { mutableStateOf(false) }
     var isSearchOpen by remember { mutableStateOf(false) }
     val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior()
     val listState = rememberLazyListState()
-    val fabVisible by remember { derivedStateOf { listState.firstVisibleItemIndex == 0 } }
-    val focusRequester = remember { FocusRequester() }
-    val scrimColor by animateColorAsState(
-        targetValue = if (fabMenuExpanded) MaterialTheme.colorScheme.scrim.copy(alpha = 0.4f) else Color.Transparent,
-        animationSpec = tween(500),
-        label = "Scrim"
-    )
 
-    Box(modifier = Modifier.fillMaxSize()) {
-        Scaffold(
-            modifier = modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
-            snackbarHost = {
-                SnackbarHost(snackbarHostState) { data ->
-                    Snackbar(
-                        snackbarData = data,
-                        containerColor = MaterialTheme.colorScheme.secondary,
-                        contentColor = MaterialTheme.colorScheme.onSecondary,
-                        actionColor = MaterialTheme.colorScheme.primary,
-                        dismissActionContentColor = MaterialTheme.colorScheme.onSecondaryContainer,
-                    )
-                }
-            },
-            topBar = {
-                LargeTopAppBar(
-                    title = {
-                        if (!isSearchOpen) Text(
-                            stringResource(R.string.app_name),
-                            style = MaterialTheme.typography.titleLarge
-                        )
-                    }, // Expressive TopAppBar Title
-                    navigationIcon = {
-                        IconButton(onClick = onNavigationIconClick) {
-                            Icon(
-                                Icons.Default.Menu,
-                                contentDescription = stringResource(R.string.navigation_drawer_open)
-                            )
-                        }
-                    },
-                    actions = {
-                        if (isSearchOpen) {
-                            TextField(
-                                value = searchQuery,
-                                onValueChange = onSearchQueryChanged,
-                                modifier = Modifier
-                                    .weight(1f)
-                                    .focusRequester(searchFocusRequester),
-                                placeholder = { Text(stringResource(R.string.search_decks)) },
-                                trailingIcon = {
-                                    IconButton(onClick = {
-                                        onSearchQueryChanged("")
-                                        isSearchOpen = false
-                                    }) {
-                                        Icon(
-                                            Icons.Default.Close,
-                                            contentDescription = stringResource(R.string.close)
-                                        )
-                                    }
-                                },
-                            )
-                        } else {
-                            IconButton(onClick = { isSearchOpen = true }) {
-                                Icon(
-                                    Icons.Default.Search,
-                                    contentDescription = stringResource(R.string.search_decks)
-                                )
-                            }
-                        }
-                    },
-                    colors = TopAppBarDefaults.topAppBarColors(
-                        containerColor = MaterialTheme.colorScheme.surface,
-                        scrolledContainerColor = MaterialTheme.colorScheme.surface,
-                        navigationIconContentColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                        actionIconContentColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                    ),
-                    scrollBehavior = scrollBehavior,
+    Scaffold(
+        modifier = modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
+        snackbarHost = {
+            SnackbarHost(snackbarHostState) { data ->
+                Snackbar(
+                    snackbarData = data,
+                    containerColor = MaterialTheme.colorScheme.secondary,
+                    contentColor = MaterialTheme.colorScheme.onSecondary,
+                    actionColor = MaterialTheme.colorScheme.primary,
+                    dismissActionContentColor = MaterialTheme.colorScheme.onSecondaryContainer,
                 )
-            },
-        ) { paddingValues ->
-            DeckPickerContent(
-                decks = decks,
-                isRefreshing = isRefreshing,
-                onRefresh = onRefresh,
-                backgroundImage = backgroundImage,
-                onDeckClick = onDeckClick,
-                onExpandClick = onExpandClick,
-                onDeckOptions = onDeckOptions,
-                onRename = onRename,
-                onExport = onExport,
-                onDelete = onDelete,
-                onRebuild = onRebuild,
-                onEmpty = onEmpty,
-                listState = listState,
-                modifier = Modifier.padding(paddingValues)
-            )
-        }
-        if (fabMenuExpanded) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(scrimColor)
-                    .clickable(
-                        interactionSource = remember { MutableInteractionSource() },
-                        indication = null,
-                        onClick = { fabMenuExpanded = false })
-            )
-        }
-        BackHandler(fabMenuExpanded) { fabMenuExpanded = false }
-
-        val onMenuItemClick = { action: () -> Unit ->
-            {
-                action()
-                fabMenuExpanded = false
             }
-        }
-
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(end = 8.dp, bottom = 32.dp),
-            contentAlignment = Alignment.BottomEnd
-        ) {
-            FloatingActionButtonMenu(
-                expanded = fabMenuExpanded,
-                button = {
-                    val fabMenuExpandedStateDescription = stringResource(R.string.fab_menu_expanded)
-                    val fabMenuCollapsedStateDescription =
-                        stringResource(R.string.fab_menu_collapsed)
-                    val fabMenuToggleContentDescription = stringResource(R.string.fab_menu_toggle)
-                    ToggleFloatingActionButton(modifier = Modifier
-                        .semantics {
-                            traversalIndex = -1f
-                            stateDescription =
-                                if (fabMenuExpanded) fabMenuExpandedStateDescription else fabMenuCollapsedStateDescription
-                            contentDescription = fabMenuToggleContentDescription
-                        }
-                        .animateFloatingActionButton(
-                            visible = fabVisible || fabMenuExpanded,
-                            alignment = Alignment.BottomEnd,
-                        )
-                        .focusRequester(focusRequester),
-                        checked = fabMenuExpanded,
-                        onCheckedChange = { fabMenuExpanded = !fabMenuExpanded }) {
-                        val imageVector by remember {
-                            derivedStateOf {
-                                if (checkedProgress > 0.5f) Icons.Filled.Close else Icons.Filled.Add
-                            }
-                        }
+        },
+        topBar = {
+            LargeTopAppBar(
+                title = {
+                    if (!isSearchOpen) Text(
+                        stringResource(R.string.app_name),
+                        style = MaterialTheme.typography.titleLarge
+                    )
+                }, // Expressive TopAppBar Title
+                navigationIcon = {
+                    IconButton(onClick = onNavigationIconClick) {
                         Icon(
-                            painter = rememberVectorPainter(imageVector),
-                            contentDescription = null,
-                            modifier = Modifier.animateIcon({ checkedProgress }),
+                            Icons.Default.Menu,
+                            contentDescription = stringResource(R.string.navigation_drawer_open)
                         )
                     }
                 },
-            ) {
-                FloatingActionButtonMenuItem(
-                    onClick = onMenuItemClick(onAddSharedDeck),
-                    icon = { Icon(Icons.Filled.Download, contentDescription = null) },
-                    text = { Text(text = stringResource(R.string.get_shared)) },
-                )
-                FloatingActionButtonMenuItem(
-                    onClick = onMenuItemClick(onAddFilteredDeck),
-                    icon = {
-                        Icon(
-                            painterResource(id = R.drawable.ic_add_filtered_deck),
-                            contentDescription = null
+                actions = {
+                    if (isSearchOpen) {
+                        TextField(
+                            value = searchQuery,
+                            onValueChange = onSearchQueryChanged,
+                            modifier = Modifier
+                                .weight(1f)
+                                .focusRequester(searchFocusRequester),
+                            placeholder = { Text(stringResource(R.string.search_decks)) },
+                            trailingIcon = {
+                                IconButton(onClick = {
+                                    onSearchQueryChanged("")
+                                    isSearchOpen = false
+                                }) {
+                                    Icon(
+                                        Icons.Default.Close,
+                                        contentDescription = stringResource(R.string.close)
+                                    )
+                                }
+                            },
                         )
-                    },
-                    text = { Text(text = stringResource(R.string.new_dynamic_deck)) },
-                )
-                FloatingActionButtonMenuItem(
-                    onClick = onMenuItemClick(onAddDeck),
-                    icon = {
-                        Icon(
-                            painterResource(id = R.drawable.ic_add_deck_filled),
-                            contentDescription = null
-                        )
-                    },
-                    text = { Text(text = stringResource(R.string.new_deck)) },
-                )
-                FloatingActionButtonMenuItem(
-                    onClick = onMenuItemClick(onAddNote),
-                    icon = {
-                        Icon(
-                            painterResource(id = R.drawable.ic_add_note),
-                            contentDescription = null
-                        )
-                    },
-                    text = { Text(text = stringResource(R.string.add_card)) },
-                )
-            }
-
+                    } else {
+                        IconButton(onClick = { isSearchOpen = true }) {
+                            Icon(
+                                Icons.Default.Search,
+                                contentDescription = stringResource(R.string.search_decks)
+                            )
+                        }
+                    }
+                },
+                colors = TopAppBarDefaults.topAppBarColors(
+                    containerColor = MaterialTheme.colorScheme.surface,
+                    scrolledContainerColor = MaterialTheme.colorScheme.surface,
+                    navigationIconContentColor = MaterialTheme.colorScheme.onSurfaceVariant,
+                    actionIconContentColor = MaterialTheme.colorScheme.onSurfaceVariant,
+                ),
+                scrollBehavior = scrollBehavior,
+            )
+        },
+        floatingActionButton = {
+            ExpandableFab(
+                onAddNote = onAddNote,
+                onAddDeck = onAddDeck,
+                onAddSharedDeck = onAddSharedDeck,
+                onAddFilteredDeck = onAddFilteredDeck
+            )
         }
+    ) { paddingValues ->
+        DeckPickerContent(
+            decks = decks,
+            isRefreshing = isRefreshing,
+            onRefresh = onRefresh,
+            backgroundImage = backgroundImage,
+            onDeckClick = onDeckClick,
+            onExpandClick = onExpandClick,
+            onDeckOptions = onDeckOptions,
+            onRename = onRename,
+            onExport = onExport,
+            onDelete = onDelete,
+            onRebuild = onRebuild,
+            onEmpty = onEmpty,
+            listState = listState,
+            modifier = Modifier.padding(paddingValues)
+        )
     }
 }
